+++ src/Dephpug/Dephpugger.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Dephpug;
+
+class Dephpugger
+{
+    public static $VERSION = '0.0.1';
+}